windows: compositor now COMPILES clean for windows (0 errors) -- reaches link
aarch64-windows-gnu compositor: 135 -> 0 compile errors. linux/macOS green.
The build now proceeds to the link stage with 102 undefined POSIX symbols
(the substrate, next). x86_64-windows also compiles bar the lua setjmp asm
(now arch-gated) + the afsrv_* frameserver exes (to be gated like a12/net).

LLP64 residual (per-site): off_t alias -> i64 on windows; lua PRNG (lmathlib
Rand64/state/rotl/nextrand/I2d) c_ulong -> usize; llex hex accumulator; lapi
GCdebt isize + 1<<40 constant; struct_nonblock_io.ofs -> i64 (SHARED
shmif_types.zig); printf-vararg c_long/c_ulong -> *longlong; ma_alsa
uframes/sframes + nfds LLP64 gates.

fd-model: cross-platform statPath/statFd (CRT _stat64/_fstat64) in
shmif_types (SHARED) routing arcan_ttf font-stat (bug 0125 EBADF panic
preserved via the identical std.posix branch); AT.FDCWD removed. std.posix.fd_t
= -1 fields -> c_int (vk_wsi/vk_gbm_kms/video). video fd dup -> CRT _dup.

comptime-prune of posix-only paths on windows: dmabuf-ino (vk x2), drm close +
evdev/psep device-stat, nbio FIFO open (mkfifo), probe blob write, fdscan
(getrlimit+linux POLL), vk drm render-node open (std.c.open void-param).

build.zig: a12/arcan-net/arcan-net-session default off on windows; don't link
posix system libs (m/rt/dl/pthread/atomic/util) on windows -- base is
kernel32/msvcrt + the (pending) substrate.

+++ src/lua54/lapi.zig
@@ -2759,11 +2759,11 @@ pub export fn lua_gc(L: [*c]lua_State, what: c_int, ...) callconv(.c) c_int {
 }
 // GC control — direct access to GCdebt for stopping/restarting GC.
 // Workaround for lua_gc being a no-op stub.
-var gc_saved_debt: c_long = 0;
+var gc_saved_debt: isize = 0;
 pub export fn gc_control_stop() void {
     if (gc_state_ptr) |g| {
         gc_saved_debt = g.*.GCdebt;
-        g.*.GCdebt = -@as(c_long, 1 << 40); // huge negative = never trigger
+        g.*.GCdebt = -@as(isize, 1 << 40); // huge negative = never trigger
     }
 }
 pub export fn gc_control_restart() void {
